pyproject.toml
[project]
name = "eroxii-alpr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "celery>=5.6.3",
    "celery-redbeat>=2.3.3",
    "fastapi[standard]>=0.136.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-decouple>=3.8",
    "python-multipart>=0.0.27",
    "pytz>=2026.1.post1",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "starlette>=1.0.0",
    "user-agents>=2.2.0",
    "uvicorn>=0.46.0",
    "weasyprint>=68.1",
    "fastmcp>=2.0.0",
    "langchain>=1.0.0",
    "langchain-huggingface>=1.2.2",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "python-telegram-bot>=21.0",
    "transformers>=5.0.0",
    "torch>=2.3.0",
    "accelerate>=0.30.0",
    "huggingface-hub>=0.23.0",
    "Pillow>=10.0.0",
    "torchvision>=0.26.0",
    "langchain-ollama>=1.1.0",
]
